部分版本更新

+++ xj-base/xj-base-admin/src/main/java/io/github/wslxm/springbootplus2/client/xj/controller/UXjAdminMsgController.java
@@ -44,7 +44,7 @@ public class UXjAdminMsgController extends BaseController<XjAdminMsgService> {
 
     @ApiOperation(value = "查询未读数量(当前登录用户)")
     @GetMapping(value = "/findUnreadNum")
-    public R<Integer> findUnreadNum() {
+    public R<Long> findUnreadNum() {
         return R.successFind(baseService.findUnreadNum());
     }
 }
