+++ xj-base/xj-base-admin/src/main/java/io/github/wslxm/springbootplus2/manage/xj/service/XjAdminMsgService.java
@@ -55,7 +55,7 @@ public interface XjAdminMsgService extends IService<XjAdminMsg> {
      * @return java.lang.Integer
      * @version 1.0.0
      */
-    Integer findUnreadNum();
+    Long findUnreadNum();
 
 
     /**
